Add target platform support (#99)
These revisions implement the new `target platform` feature. With this feature, test implementers are able to assign each test to a specific "platform", which is a scenario-defined set of execution characteristics. For example, you might define a set of **format** targets - `SMALL` for mobile format, `MEDIUM` for compact format, and `LARGE` for desktop format. The implementation of the `activatePlatform` method would set the browser dimensions and add or remove flags that control mobile-format rendering.