Ignore shadow DOM failures; Reformat example page (#187)
Prior to this update, implementation to ignore shadow DOM failures in the JUnit model tests was missing. These tests are currently disabled, but I turn them back on occasionally and may eventually switch to JUnit as the primary framework for the unit tests. These failures now trigger assumption violations, which is the correct JUnit idiom.
In addition to these test fixes, I also made the following revisions:
* Reformat the Example page to make it more tidy.
* For Selenium 2 (Java 7), upgrade to HtmlUnitDriver 2.24 and remove the direct dependency on HtmlUnit itself.
* Add debug logging to report actions and outcomes of Selenium Grid modeling process.